# Perform linear algebra using faster BLAS implementation, if possible