Fixes #593, @rafaelha, @theodotk

Here's my abbreviated performance test. It would be nice to write a set of `.npy` files (with the result `C = A @ B`) before the change and compare them to after the change. I want to make sure no arithmetic is broken.

### Script

```python
import galois

dim1 = 10

for p in [2, 7, 1_009, 1_000_003]:
    GF = galois.GF(p)

    for dim2 in [10, 100, 1_000, 10_000, 100_000, 1_000_000]:
        A = GF.Random((dim1, dim2))
        B = GF.Random((dim2, dim1))
        print(p, dim1, dim2)
        %timeit A @ B
```

### Before

```
2 10 10
29.7 μs ± 718 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
2 10 100
37.1 μs ± 1.28 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
2 10 1000
90.3 μs ± 1.88 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
2 10 10000
599 μs ± 8.49 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
2 10 100000
6.47 ms ± 163 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
2 10 1000000
220 ms ± 15.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
7 10 10
30.3 μs ± 782 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 100
38 μs ± 926 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 1000
89.7 μs ± 2.27 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 10000
635 μs ± 7.27 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
7 10 100000
6.5 ms ± 135 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
7 10 1000000
211 ms ± 8.39 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
1009 10 10
30.2 μs ± 679 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1009 10 100
37.4 μs ± 998 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1009 10 1000
92.1 μs ± 1.06 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1009 10 10000
976 μs ± 6.54 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
1009 10 100000
11 ms ± 158 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
1009 10 1000000
374 ms ± 11.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
1000003 10 10
31.5 μs ± 1.43 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1000003 10 100
38.2 μs ± 528 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1000003 10 1000
107 μs ± 1.21 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1000003 10 10000
976 μs ± 14.1 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
1000003 10 100000
11 ms ± 126 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
1000003 10 1000000
373 ms ± 14.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### After

By my eye, the change is always faster than before.

```
2 10 10
23.6 μs ± 809 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
2 10 100
25.9 μs ± 515 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
2 10 1000
39.4 μs ± 324 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
2 10 10000
157 μs ± 2.83 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
2 10 100000
1.36 ms ± 71.3 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
2 10 1000000
21.9 ms ± 1.16 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
7 10 10
23.9 μs ± 379 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 100
26.3 μs ± 239 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 1000
39.7 μs ± 372 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 10000
160 μs ± 3.41 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 100000
1.34 ms ± 34.6 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
7 10 1000000
34.5 ms ± 780 μs per loop (mean ± std. dev. of 7 runs, 10 loops each)
1009 10 10
24.3 μs ± 53.6 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1009 10 100
27.9 μs ± 361 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1009 10 1000
43.6 μs ± 1.19 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1009 10 10000
221 μs ± 8.66 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
1009 10 100000
2.8 ms ± 105 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
1009 10 1000000
35.2 ms ± 1.01 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
1000003 10 10
26.4 μs ± 421 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1000003 10 100
29.3 μs ± 249 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1000003 10 1000
46.9 μs ± 400 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
1000003 10 10000
958 μs ± 13.5 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
1000003 10 100000
11.2 ms ± 327 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
1000003 10 1000000
361 ms ± 13.7 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)

```