@mhostetter 2025-04-09
Here's another test. It demonstrates that `np.float32` (when available) is faster than `np.float64`, `np.int64`, and `np.object_`.

```python
import galois

dim1 = 10

for p in [7,]:
    GF = galois.GF(p)

    for dim2 in [10, 100, 1_000, 10_000, 100_000, 1_000_000]:
        A = GF.Random((dim1, dim2))
        B = GF.Random((dim2, dim1))
        print(p, dim1, dim2)
        %timeit A @ B
```

### Uses `np.float32`

```python
    if max_value <= 2 ** (np.finfo(np.float32).nmant + 1):
        dtype = np.float32
    elif max_value <= 2 ** (np.finfo(np.float64).nmant + 1):
        dtype = np.float64
    elif max_value <= np.iinfo(np.int64).max:
        dtype = np.int64
    else:
        dtype = np.object_
```

```
7 10 10
25.1 μs ± 1.65 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 100
27.4 μs ± 525 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 1000
40.5 μs ± 1.01 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 10000
160 μs ± 2.04 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 100000
1.3 ms ± 32.1 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
7 10 1000000
36 ms ± 2.85 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### Uses `np.float64`

```python
    if max_value <= 2 ** (np.finfo(np.float64).nmant + 1):
        dtype = np.float64
    elif max_value <= np.iinfo(np.int64).max:
        dtype = np.int64
    else:
        dtype = np.object_
```

```
7 10 10
25.6 μs ± 576 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 100
28.8 μs ± 2.47 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 1000
43.9 μs ± 1.67 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 10000
217 μs ± 9.11 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
7 10 100000
3.78 ms ± 219 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
7 10 1000000
35.6 ms ± 1.22 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### Uses `np.int64`

```python
    if max_value <= np.iinfo(np.int64).max:
        dtype = np.int64
    else:
        dtype = np.object_
```

```
7 10 10
23.1 μs ± 619 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 100
29.9 μs ± 781 ns per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 1000
100 μs ± 2.16 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 10000
975 μs ± 8.55 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
7 10 100000
11.4 ms ± 169 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
7 10 1000000
362 ms ± 12.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### Uses `np.object_`

```python
    dtype = np.object_
```

```
7 10 10
41.1 μs ± 1.26 μs per loop (mean ± std. dev. of 7 runs, 10,000 loops each)
7 10 100
247 μs ± 2.95 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
7 10 1000
2.31 ms ± 33.9 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
7 10 10000
22.3 ms ± 299 μs per loop (mean ± std. dev. of 7 runs, 10 loops each)
7 10 100000
218 ms ± 2.31 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
7 10 1000000
2.18 s ± 19 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```