@mhostetter 2025-04-10
I think I'm probably good on performance tests. I want to add the convert to int before modulo and see if I can realize the performance benefit, though.