@mhostetter 2025-04-12
To test `np.int64` vs `np.uint64`. This size uses that dtype in the `if` statement. It does appear `uint64` is slower than `int64`.

```python
import galois

dim1 = 10

for p in [1_000_003,]:
    GF = galois.GF(p)

    for dim2 in [1_000_000]:
        A = GF.Random((dim1, dim2))
        B = GF.Random((dim2, dim1))
        print(p, dim1, dim2)
        %timeit A @ B
```

### `np.int64` (before)

```
1000003 10 1000000
363 ms ± 10.9 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### `np.uint64` (after)

```
1000003 10 1000000
379 ms ± 22.4 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```

### `np.object_`

```
1000003 10 1000000
4.09 s ± 43.8 ms per loop (mean ± std. dev. of 7 runs, 1 loop each)
```