@mhostetter 2025-04-12
To test floating-point modulo or integer modulo. I'm not really seeing a performance benefit @rafaelha. I'm going to hold off on that change for now, but am opening to adding it if you can show evidence of its utility.

```python
import galois

dim1 = 10

for p in [1_009,]:
    GF = galois.GF(p)

    for dim2 in [10_000, 100_000, 1_000_000]:
        A = GF.Random((dim1, dim2))
        B = GF.Random((dim2, dim1))
        print(p, dim1, dim2)
        %timeit A @ B
```

### Floating-point modulo

```python
    # Reduce the result mod p
    # if max_value <= np.iinfo(np.int64).max:
    #     # Performing the modulo operation on an int is faster than a float. Convert back to int64 for the time
    #     # savings, if possible.
    #     c = c.astype(np.int64)
    c = c % field.characteristic
```

```
1009 10 10000
213 μs ± 3.92 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
1009 10 100000
4.04 ms ± 264 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
1009 10 1000000
35.8 ms ± 1.1 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```

### Integer modulo

```python
    # Reduce the result mod p
    if max_value <= np.iinfo(np.int64).max:
        # Performing the modulo operation on an int is faster than a float. Convert back to int64 for the time
        # savings, if possible.
        c = c.astype(np.int64)
    c = c % field.characteristic
```

```
1009 10 10000
227 μs ± 6.44 μs per loop (mean ± std. dev. of 7 runs, 1,000 loops each)
1009 10 100000
4.32 ms ± 204 μs per loop (mean ± std. dev. of 7 runs, 100 loops each)
1009 10 1000000
36.9 ms ± 1.34 ms per loop (mean ± std. dev. of 7 runs, 10 loops each)
```