@mhostetter 2025-04-12
Using `np.uint64` results in the some errors in the test suite. I can't reproduce on my machine, but I consistently get in CI. There's minimal performance benefit, so I'm going to omit that rung of the ladder.